refactor: Extract error handling to errors.go module
- Create new errors.go file with error types and handlers
- Move ErrorType, CPUError, ErrorHandler, StrictErrorHandler, and LoggingErrorHandler from cpu.go
- Maintain all existing functionality and API compatibility
- All tests pass successfully
